+++ src/main/kotlin/no/risc/risc/models/RiScWrapperObject.kt
@@ -1,7 +1,5 @@
 package no.risc.risc.models
 
-import no.risc.infra.connector.UserInfo
-
 data class RiScWrapperObject(
     val riSc: String,
     val isRequiresNewApproval: Boolean,
