+++ src/main/kotlin/no/risc/risc/models/UserInfo.kt
@@ -1,0 +1,11 @@
+package no.risc.risc.models
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties
+
+@JsonIgnoreProperties(ignoreUnknown = true)
+data class UserInfo(
+    val name: String,
+    val email: String,
+) {
+    fun isValid(): Boolean = name.isNotBlank() && email.isNotBlank()
+}
